Refactor modal to standalone component
